pyproject.toml
[tool.poetry]
name = "swarm-devteam"
version = "0.1.0"
description = ""
authors = ["Marcin Adamowski"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
swarm = {git = "ssh://git@github.com/openai/swarm.git", develop = true}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

package-mode = false
